Added ugly solution to 2017 - Haskell - Day19.
This solution is a bit ugly: it turns out that Data.Graph does not support
adding value to edges, which means my graph-creating function has to return an
additional function that does a mapping edge -> cost by doing a lookup in a
hidden map.

On top of it, the Node and NodeMap types that are used to do the accumulation
are not exactly representative of the data we end up needing, meaning there's a
lot of inelegant transformations to do (see getNode and getEdgeCost for
instance).

Finally, this solution relies on an undocumented property of the "reachable"
function: the fact that it is going to enumerate nodes in the "correct" order.

But hey; at least it works. :)