Added horrible 2017 - Haskell - Day23 solution.
I have NOT understood yet how to do programmatical analysis of the input. I
added support for comments in the input, and commented it until I understood
what it was doing, and recoded the same code in pure Haskell.

As a result, my solution is specific to my input, which seems wrong.