2024 Brainfuck Day 14 part 1
It requires modifying constants to work on the test input, but it DOES
work on the test input. Untested (for now) on a real input.

Since the real result will not fit in an int32, the result is printed
as a multiplication of int32 which is left as an exercise to the
reader.